Move Prometheus stuff
