+++ src/Promitor.Agents.Scraper/AzureMonitorClientFactory.cs
@@ -4,8 +4,8 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Promitor.Core;
+using Promitor.Core.Metrics;
 using Promitor.Core.Scraping.Configuration.Model;
-using Promitor.Core.Telemetry.Metrics.Interfaces;
 using Promitor.Integrations.AzureMonitor;
 using Promitor.Integrations.AzureMonitor.Configuration;
