+++ src/Promitor.Core.Scraping/Factories/MetricScraperFactory.cs
@@ -4,9 +4,8 @@ using Microsoft.Extensions.Logging;
 using Promitor.Core.Scraping.Configuration.Model;
 using Promitor.Core.Scraping.Configuration.Model.Metrics;
 using Promitor.Core.Scraping.Interfaces;
-using Promitor.Core.Scraping.Prometheus.Interfaces;
+using Promitor.Core.Scraping.Metrics.Sinks;
 using Promitor.Core.Scraping.ResourceTypes;
-using Promitor.Core.Scraping.Sinks;
 using Promitor.Integrations.AzureMonitor;
 
 namespace Promitor.Core.Scraping.Factories
