+++ src/Promitor.Core.Scraping/Metrics/Sinks/MetricSinkType.cs
@@ -1,4 +1,4 @@
-﻿namespace Promitor.Core.Scraping.Sinks
+﻿namespace Promitor.Core.Scraping.Metrics.Sinks
 {
     public enum MetricSinkType
     {
