+++ src/Promitor.Core.Scraping/Scraper.cs
@@ -6,8 +6,7 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Promitor.Core.Scraping.Configuration.Model.Metrics;
 using Promitor.Core.Scraping.Interfaces;
-using Promitor.Core.Scraping.Prometheus.Interfaces;
-using Promitor.Core.Scraping.Sinks;
+using Promitor.Core.Scraping.Metrics.Sinks;
 using Promitor.Integrations.AzureMonitor;
 
 namespace Promitor.Core.Scraping
