+++ src/Promitor.Core.Scraping/ScraperConfiguration.cs
@@ -1,7 +1,7 @@
 ﻿using GuardNet;
 using Microsoft.Extensions.Logging;
-using Promitor.Core.Scraping.Prometheus.Interfaces;
-using Promitor.Core.Scraping.Sinks;
+using Promitor.Core.Scraping.Interfaces;
+using Promitor.Core.Scraping.Metrics.Sinks;
 using Promitor.Integrations.AzureMonitor;
 
 namespace Promitor.Core.Scraping
