+++ src/Promitor.Tests.Unit/Generators/Config/RuntimeConfigurationGenerator.cs
@@ -10,7 +10,7 @@ using Promitor.Agents.Core.Configuration.Telemetry;
 using Promitor.Agents.Core.Configuration.Telemetry.Sinks;
 using Promitor.Agents.Scraper.Configuration;
 using Promitor.Core.Scraping.Configuration.Runtime;
-using Promitor.Core.Telemetry.Metrics.Configuration;
+using Promitor.Integrations.Sinks.Prometheus.Configuration;
 
 namespace Promitor.Tests.Unit.Generators.Config
 {
